Move solution to separate file
